now processGeometry node works in WebGL

+++ src/webgpu/shaderity_shaders/nodes/AttributeJoint.vert.wgsl
@@ -1,3 +1,3 @@
-fn attributeJoint(outValue: ptr<function, vec4<f32>>) {
+fn attributeJoint(outValue: ptr<function, vec4<u32>>) {
   *outValue = a_joint;
 }
